src/knmi_dataset_downloader/knmi_dataset_api/v1/v1_request_builder.py
from __future__ import annotations
from kiota_abstractions.base_request_builder import BaseRequestBuilder
from kiota_abstractions.get_path_parameters import get_path_parameters
from kiota_abstractions.request_adapter import RequestAdapter
from typing import Any, Callable, Dict, List, Optional, TYPE_CHECKING, Union

if TYPE_CHECKING:
    from .datasets.datasets_request_builder import DatasetsRequestBuilder

class V1RequestBuilder(BaseRequestBuilder):
    """
    Builds and executes requests for operations under /v1
    """
    def __init__(self,request_adapter: RequestAdapter, path_parameters: Union[str, Dict[str, Any]]) -> None:
        """
        Instantiates a new V1RequestBuilder and sets the default values.
        param path_parameters: The raw url or the url-template parameters for the request.
        param request_adapter: The request adapter to use to execute the requests.
        Returns: None
        """
        super().__init__(request_adapter, "{+baseurl}/v1", path_parameters)
    
    @property
    def datasets(self) -> DatasetsRequestBuilder:
        """
        The datasets property
        """
        from .datasets.datasets_request_builder import DatasetsRequestBuilder

        return DatasetsRequestBuilder(self.request_adapter, self.path_parameters)
    

